[#258] Refactor clean architecture
