+++ bricks/template/__brick__/{{project_name.snakeCase()}}/lib/app/resources/app_colors.dart
