+++ bricks/template/__brick__/{{project_name.snakeCase()}}/lib/app/screens/home/home_view_state.dart
