+++ bricks/template/__brick__/{{project_name.snakeCase()}}/lib/data/datasources/remote/response/user_response.dart
@@ -1,4 +1,5 @@
 import 'package:json_annotation/json_annotation.dart';
+import 'package:{{project_name.snakeCase()}}/domain/models/user.dart';
 
 part 'user_response.g.dart';
 
@@ -13,4 +14,9 @@ class UserResponse {
       _$UserResponseFromJson(json);
 
   Map<String, dynamic> toJson() => _$UserResponseToJson(this);
+
+  User toUser() => User(
+    email: email,
+    username: username,
+  );
 }
