+++ bricks/template/__brick__/{{project_name.snakeCase()}}/lib/domain/repositories/credential_repository.dart
@@ -1,0 +1,5 @@
+import 'package:{{project_name.snakeCase()}}/domain/models/user.dart';
+
+abstract class CredentialRepository {
+  Future<List<User>> getUsers();
+}
