+++ bricks/template/__brick__/{{project_name.snakeCase()}}/lib/domain/usecases/base/base_use_case.dart
