+++ bricks/template/__brick__/{{project_name.snakeCase()}}/lib/domain/usecases/base/use_case_result.dart
