+++ bricks/template/__brick__/{{project_name.snakeCase()}}/test/mocks/generate_mocks.dart
@@ -1,7 +1,7 @@
 import 'package:dio/dio.dart';
-import 'package:{{project_name.snakeCase()}}/api/api_service.dart';
-import 'package:{{project_name.snakeCase()}}/api/repository/credential_repository.dart';
-import 'package:{{project_name.snakeCase()}}/usecases/user/get_users_use_case.dart';
+import 'package:{{project_name.snakeCase()}}/data/datasources/remote/api_service.dart';
+import 'package:{{project_name.snakeCase()}}/domain/repositories/credential_repository.dart';
+import 'package:{{project_name.snakeCase()}}/domain/usecases/user/get_users_use_case.dart';
 import 'package:mockito/annotations.dart';
 
 @GenerateMocks([
