+++ bricks/template/__brick__/{{project_name.snakeCase()}}/test/usecases/user/get_users_use_case_test.dart
@@ -1,8 +1,8 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
-import 'package:{{project_name.snakeCase()}}/usecases/base/base_use_case.dart';
-import 'package:{{project_name.snakeCase()}}/usecases/user/get_users_use_case.dart';
-import 'package:{{project_name.snakeCase()}}/model/user.dart';
+import 'package:{{project_name.snakeCase()}}/domain/usecases/base/base_use_case.dart';
+import 'package:{{project_name.snakeCase()}}/domain/usecases/user/get_users_use_case.dart';
+import 'package:{{project_name.snakeCase()}}/domain/models/user.dart';
 
 import '../../mocks/generate_mocks.mocks.dart';
 import '../../mocks/response/user_response_mocks.dart';
