+++ bricks/template/__brick__/{{project_name.snakeCase()}}/test/viewmodel/home_view_model_test.dart
@@ -1,10 +1,10 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
-import 'package:{{project_name.snakeCase()}}/usecases/base/base_use_case.dart';
-import 'package:{{project_name.snakeCase()}}/home_view_model.dart';
+import 'package:{{project_name.snakeCase()}}/app/screens/home/home_view_model.dart';
+import 'package:{{project_name.snakeCase()}}/domain/usecases/base/base_use_case.dart';
+import 'package:{{project_name.snakeCase()}}/domain/models/user.dart';
 import 'package:{{project_name.snakeCase()}}/main.dart';
-import 'package:{{project_name.snakeCase()}}/model/user.dart';
 
 import '../mocks/generate_mocks.mocks.dart';
 import '../mocks/response/user_response_mocks.dart';
